refactoring: tabs component  - component  - docs
