+++ src/components/FwbTabs/injection/config.ts
@@ -1,6 +1,6 @@
 // dependency injection config for tabs
 
-export const TAB_STYLE_INJECTION_KEY = 'flowbite-tab-style-injection'
-export const TAB_ACTIVE_NAME_INJECTION_KEY = 'flowbite-tab-active-name-injection'
-export const TAB_VISIBILITY_DIRECTIVE_INJECTION_KEY = 'flowbite-tab-visibility-directive-injection'
 export const TAB_ACTIVATE_INJECTION_KEY = 'flowbite-tab-activate-func-injection'
+export const TAB_ACTIVE_NAME_INJECTION_KEY = 'flowbite-tab-active-name-injection'
+export const TAB_STYLE_INJECTION_KEY = 'flowbite-tab-style-injection'
+export const TAB_VISIBILITY_DIRECTIVE_INJECTION_KEY = 'flowbite-tab-visibility-directive-injection'
