+++ src/components/FwbTabs/types.ts
@@ -1,0 +1,1 @@
+export type TabsVariant = 'default' | 'underline' | 'pills'
