--- src/components/Tabs/types.ts
@@ -1,1 +1,0 @@
-export type TabsVariant = 'default' | 'underline' | 'pills'
