+++ src/index.ts
@@ -37,9 +37,9 @@ export { default as FwbTableCell } from './components/FwbTable/FwbTableCell.vue'
 export { default as FwbTableHead } from './components/FwbTable/FwbTableHead.vue'
 export { default as FwbTableHeadCell } from './components/FwbTable/FwbTableHeadCell.vue'
 export { default as FwbTableRow } from './components/FwbTable/FwbTableRow.vue'
+export { default as FwbTabs } from './components/FwbTabs/FwbTabs.vue'
+export { default as FwbTab } from './components/FwbTabs/FwbTab.vue'
 
-export { default as Tabs } from './components/Tabs/Tabs.vue'
-export { default as Tab } from './components/Tabs/components/Tab/Tab.vue'
 export { default as FlowbiteThemable } from './components/utils/FlowbiteThemable/FlowbiteThemable.vue'
 export { default as FlowbiteThemableChild } from './components/utils/FlowbiteThemable/components/FlowbiteThemableChild/FlowbiteThemableChild.vue'
 export { default as Sidebar } from './components/Sidebar/Sidebar.vue'
